extract file data-provider out of the file-list component
